+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### Unreleased
  * Fixed issue with mounting SPContentDatabase
  * Fixed issue with SPShellAdmin and Content Database method
- * Added .Net 4.6 support check
+ * Added .Net 4.6 support check to SPInstall and SPInstallPrereqs
  * Improved code styling
 
 ### 1.3